requirements.txt
numpy<2.0
pyniryo2
pyyaml
